PR 생성 시 서버 API 호출로 브랜치커밋 가이드 자동 댓글 : refactor : 골격 및 전체적인 프로젝트 구조 리펙토링 https://github.com/Cassiiopeia/suh-project-utility/issues/48
