+++ Suh-Common/src/main/java/me/suhsaechan/common/deprecated/LogMonitoringInvocation.java
