+++ Suh-Common/src/main/java/me/suhsaechan/common/deprecated/LogUtil.java
