+++ Suh-Common/src/main/java/me/suhsaechan/common/service/WebDriverManager.java
@@ -1,4 +1,4 @@
-package me.suhsaechan.web.config;
+package me.suhsaechan.common.service;
 
 import java.net.MalformedURLException;
 import java.net.URL;
