+++ Suh-Domain-Docker/src/main/java/me/suhsaechan/docker/service/DockerLogService.java
@@ -19,8 +19,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.suhsaechan.common.config.SshConnectionProperties;
 import me.suhsaechan.common.util.SshCommandExecutor;
+import me.suhsaechan.common.properties.SshConnectionProperties;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
