+++ Suh-Domain-Github/src/main/java/me/suhsaechan/github/repository/GithubIssueHelperRepository.java
@@ -1,8 +1,8 @@
-package me.suhsaechan.common.repository;
+package me.suhsaechan.github.repository;
 
 import java.util.Optional;
 import java.util.UUID;
-import me.suhsaechan.common.entity.GithubIssueHelper;
+import me.suhsaechan.github.entity.GithubIssueHelper;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GithubIssueHelperRepository extends JpaRepository<GithubIssueHelper, UUID> {
