+++ Suh-Domain-Github/src/main/java/me/suhsaechan/github/repository/GithubRepositoryRepository.java
@@ -1,7 +1,7 @@
-package me.suhsaechan.common.repository;
+package me.suhsaechan.github.repository;
 
 import java.util.UUID;
-import me.suhsaechan.common.entity.GithubRepository;
+import me.suhsaechan.github.entity.GithubRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GithubRepositoryRepository extends JpaRepository<GithubRepository, UUID> {
