+++ Suh-Web/src/main/java/me/suhsaechan/web/controller/api/NoticeController.java
@@ -1,8 +1,8 @@
-package me.suhsaechan.web.controller;
+package me.suhsaechan.web.controller.api;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.suhsaechan.web.config.UserAuthority;
+import me.suhsaechan.common.service.UserAuthority;
 import me.suhsaechan.notice.service.NoticeService;
 import me.suhsaechan.suhlogger.annotation.LogMonitor;
 import org.springframework.http.MediaType;
