+++ Suh-Web/src/main/java/me/suhsaechan/web/controller/view/GlobalModelController.java
@@ -1,11 +1,11 @@
-package me.suhsaechan.web.controller;
+package me.suhsaechan.web.controller.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.suhsaechan.web.config.ServerInfo;
-import me.suhsaechan.common.util.security.AESUtil;
+import me.suhsaechan.common.util.AESUtil;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ModelAttribute;
